Issue #3542: refactored check converters outside of AutomaticBean
