+++ src/test/java/com/puppycrawl/tools/checkstyle/filters/SuppressionFilterTest.java
@@ -29,6 +29,7 @@ import java.net.HttpURLConnection;
 import java.net.URI;
 import java.net.URL;
 
+import org.apache.commons.beanutils.BeanUtilsBean;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -66,7 +67,8 @@ public class SuppressionFilterTest extends AbstractModuleTestSupport {
         EqualsVerifier
                 .forClass(SuppressionFilter.class)
                 .usingGetClass()
-                .withIgnoredFields("file", "optional", "configuration")
+                .withIgnoredFields("file", "optional", "configuration", "beanUtils")
+                .withPrefabValues(BeanUtilsBean.class, new BeanUtilsBean(), new BeanUtilsBean())
                 .suppress(Warning.NONFINAL_FIELDS)
                 .verify();
     }
